Refactor the DataSkipping test files ahead of the feature-area split
Reorganizes the DataSkipping test sources (DataSkippingDeltaTests and the shared DataSkippingDeltaTestsUtils) to prepare for splitting the suites by feature area. This is a no-op refactor: no test behavior changes, only how the shared body and helpers are arranged.
